Revert "Merge branch 'master' into master"
This reverts commit ed6ae650f8a5e4c9779170ff295a091632825923, reversing
changes made to b8ec166f0ca5f8f520063f65989175514060674d.

+++ src/common/stream.cc
@@ -1,6 +1,6 @@
 /* stream.cc
    Mathieu Stefani, 05 September 2015
-
+   
 */
 
 #include <iostream>
@@ -71,7 +71,7 @@ DynamicStreamBuf::reserve(size_t size)
 
 bool
 StreamCursor::advance(size_t count) {
-    if (static_cast<ssize_t>(count) > buf->in_avail())
+    if (count > buf->in_avail())
         return false;
 
     for (size_t i = 0; i < count; ++i) {
